Moved the react components directory into the docs directory
This is to ensure the references still stay in tack when tagging versions into their own folders

OHM-934
